refactor(app): move ODD instrument pages
