+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -40,8 +40,8 @@ import { RunningProtocol } from '../pages/RunningProtocol'
 import { RunSummary } from '../pages/RunSummary'
 import { UpdateRobot } from '../pages/UpdateRobot/UpdateRobot'
 import { UpdateRobotDuringOnboarding } from '../pages/UpdateRobot/UpdateRobotDuringOnboarding'
-import { InstrumentsDashboard } from '../pages/InstrumentsDashboard'
-import { InstrumentDetail } from '../pages/InstrumentDetail'
+import { InstrumentsDashboard } from '../pages/ODD/InstrumentsDashboard'
+import { InstrumentDetail } from '../pages/ODD/InstrumentDetail'
 import { Welcome } from '../pages/Welcome'
 import { InitialLoadingScreen } from '../pages/InitialLoadingScreen'
 import { DeckConfigurationEditor } from '../pages/DeckConfiguration'
