+++ app/src/pages/ODD/InstrumentsDashboard/__tests__/PipetteRecalibrationODDWarning.test.tsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import { screen } from '@testing-library/react'
 import { describe, it } from 'vitest'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 
 import { PipetteRecalibrationODDWarning } from '../PipetteRecalibrationODDWarning'
