+++ app/src/pages/ODD/InstrumentsDashboard/index.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react'
 import { useInstrumentsQuery } from '@opentrons/react-api-client'
 import { DIRECTION_COLUMN, Flex, SPACING } from '@opentrons/components'
-import { PipetteWizardFlows } from '../../organisms/PipetteWizardFlows'
-import { Navigation } from '../../organisms/Navigation'
-import { AttachedInstrumentMountItem } from '../../organisms/InstrumentMountItem'
-import { GripperWizardFlows } from '../../organisms/GripperWizardFlows'
-import { getShowPipetteCalibrationWarning } from '../../organisms/Devices/utils'
+import { PipetteWizardFlows } from '../../../organisms/PipetteWizardFlows'
+import { Navigation } from '../../../organisms/Navigation'
+import { AttachedInstrumentMountItem } from '../../../organisms/InstrumentMountItem'
+import { GripperWizardFlows } from '../../../organisms/GripperWizardFlows'
+import { getShowPipetteCalibrationWarning } from '../../../organisms/Devices/utils'
 import { PipetteRecalibrationODDWarning } from './PipetteRecalibrationODDWarning'
 import type { GripperData, PipetteData } from '@opentrons/api-client'
